% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incusyte.R
\name{format}
\alias{format}
\title{Formats the wide format to a long format}
\usage{
format(df, formatter = format_group)
}
\arguments{
\item{df}{data frame with all observations}
}
\description{
Renames A-D and E-H to X1 - X4 to simplify the calculation
    and keeps the Group information as a separate column.
}
