% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incusyte.R
\name{format_group}
\alias{format_group}
\title{formats a group}
\usage{
format_group(df, letters, i = 1:12)
}
\arguments{
\item{df}{loaded plate}

\item{letters}{either c("A", "B", "C", "D") or c("E", "F", "G", "H")}

\item{i}{representing the number plates as a list. Default = 1:12}
}
\description{
Given a plate which is labeled A to H, when
    referring to a group we mean either A to D or E to H.
    A1 to H1 represents WHAT. Where A1 to D1 represents
    one WHAT.
}
