% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{silent_library}
\alias{silent_library}
\title{Make calls to \code{library} extra silent}
\usage{
silent_library(package, help, pos = 2, lib.loc = NULL,
  character.only = FALSE, logical.return = FALSE, warn.conflicts = FALSE,
  quietly = TRUE, verbose = FALSE)
}
\arguments{
\item{library_name}{library to load, silencing all warnings}
}
\description{
Command line tools don’t want to clutter their output with unnecessary noise.
  This replaces the default \code{library} arguments to ensure this silence. It
  additionally wraps all calls in \code{suppressMessages} to be extra silent.
  This is mainly necessary because not all packages play nicely and use
  \code{message} inappropriately instead of \code{packageStartupMessage}.
}
